Separates the 'big_query/client/job_types'
