+++ lib/big_query/client.rb
@@ -7,6 +7,7 @@ require 'big_query/client/load'
 require 'big_query/client/hashable'
 require 'big_query/client/options'
 require 'big_query/client/response'
+require 'big_query/client/job_types'
 
 module BigQuery
   class Client
@@ -19,6 +20,7 @@ module BigQuery
     include BigQuery::Client::Hashable
     include BigQuery::Client::Options
     include BigQuery::Client::Response
+    include BigQuery::Client::JobTypes
 
     attr_accessor :dataset, :project_id
